관통프로젝트/03_DataEngineeringPJT/08_PJT/kafka&flink&docker 가이드/requirements.txt
feedparser==6.0.11
kafka-python==2.0.2
apache-flink==1.20.0
beautifulsoup4==4.13.3
pydantic==2.10.6
psycopg2-binary==2.9.10
python-dotenv==1.0.1
tiktoken==0.8.0
pyspark==3.5.4
matplotlib==3.10.0
hdfs==2.7.3
elasticsearch==8.17.1
openai==2.8.1